[AIRFLOW-7040] Move tests/utils/contrib packages to tests/utils
