+++ tests/providers/google/cloud/operators/test_cloud_build_system_helper.py
@@ -24,8 +24,8 @@ import os
 from tempfile import TemporaryDirectory
 from urllib.parse import urlparse
 
-from tests.contrib.utils.logging_command_executor import LoggingCommandExecutor
 from tests.providers.google.cloud.utils.gcp_authenticator import GCP_CLOUD_BUILD_KEY, GcpAuthenticator
+from tests.utils.logging_command_executor import LoggingCommandExecutor
 
 GCE_INSTANCE = os.environ.get("GCE_INSTANCE", "testinstance")
 GCP_PROJECT_ID = os.environ.get("GCP_PROJECT_ID", "example-project")
