+++ tests/providers/google/cloud/operators/test_compute_system_helper.py
@@ -19,8 +19,8 @@
 import argparse
 import os
 
-from tests.contrib.utils.logging_command_executor import LoggingCommandExecutor
 from tests.providers.google.cloud.utils.gcp_authenticator import GCP_COMPUTE_KEY, GcpAuthenticator
+from tests.utils.logging_command_executor import LoggingCommandExecutor
 
 GCE_INSTANCE = os.environ.get('GCE_INSTANCE', 'testinstance')
 GCP_PROJECT_ID = os.environ.get('GCP_PROJECT_ID', 'example-project')
