+++ tests/providers/google/cloud/operators/test_gcs_system_helper.py
@@ -18,7 +18,7 @@
 # under the License.
 import os
 
-from tests.contrib.utils.logging_command_executor import LoggingCommandExecutor
+from tests.utils.logging_command_executor import LoggingCommandExecutor
 
 BUCKET_1 = os.environ.get("GCP_GCS_BUCKET_1", "test-gcs-example-bucket")
 BUCKET_2 = os.environ.get("GCP_GCS_BUCKET_1", "test-gcs-example-bucket-2")
