+++ tests/test_utils/gcp_system_helpers.py
@@ -24,10 +24,10 @@ import pytest
 from google.auth.environment_vars import CREDENTIALS
 
 from airflow.providers.google.cloud.utils.credentials_provider import provide_gcp_conn_and_credentials
-from tests.contrib.utils.logging_command_executor import get_executor
 from tests.providers.google.cloud.utils.gcp_authenticator import GCP_GCS_KEY
 from tests.test_utils import AIRFLOW_MAIN_FOLDER
 from tests.test_utils.system_tests_class import SystemTest
+from tests.utils.logging_command_executor import get_executor
 
 CLOUD_DAG_FOLDER = os.path.join(
     AIRFLOW_MAIN_FOLDER, "airflow", "providers", "google", "cloud", "example_dags"
