+++ tests/test_utils/system_tests_class.py
@@ -26,8 +26,8 @@ from airflow.exceptions import AirflowException
 from airflow.models.dagbag import DagBag
 from airflow.utils.file import mkdirs
 from airflow.utils.log.logging_mixin import LoggingMixin
-from tests.contrib.utils.logging_command_executor import get_executor
 from tests.test_utils import AIRFLOW_MAIN_FOLDER
+from tests.utils.logging_command_executor import get_executor
 
 DEFAULT_DAG_FOLDER = os.path.join(AIRFLOW_MAIN_FOLDER, "airflow", "example_dags")
